Api/requirements.txt
Flask
Flask-RESTful
flask_cors
requests
mariadb
argon2-cffi
PyJWT